requirements.txt
requests
beautifulsoup4
pandas
transformers
tensforflow
torch
mysql-connector-python
flask
openai
langchain
streamlit